requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
pillow
opencv-python-headless
numpy
werkzeug
